+++ CHANGES.md
@@ -6,7 +6,7 @@ List changes at major/minor level. Patches should be straightforward.
 - Stop adding breaking changes in minor versions.
 - Drop RXJS, use ol/native mechanisms instead.
 - Remove the UnByKeyAll method, you can use the native ol/Observable.unByKeyAll method instead.
-- Refactor and improve the feature.utils getCenterOfArea to geometry.utils getCenterOfGeometry.
+- Refactor and improve the feature.utils getCenterOfArea to geometry.utils getGeometryCenter.
 - Add an animation possibility on the olc view.fit.
 - Use Kebab case for file names.